backend-server/requirements.txt
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
google-api-core==1.22.1
google-api-python-client==1.11.0
google-auth==1.21.0
google-auth-httplib2==0.0.4
google-cloud==0.34.0
google-cloud-core==1.4.1
google-cloud-texttospeech==2.2.0
google-cloud-translate==3.0.1
google-cloud-vision==1.0.0
googleapis-common-protos==1.52.0
grpcio==1.31.0
httplib2==0.18.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
libcst==0.3.10
MarkupSafe==1.1.1
mypy-extensions==0.4.3
numpy==1.19.1
pandas==1.1.1
playsound==1.2.2
proto-plus==1.8.1
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydub==0.24.1
python-dateutil==2.8.1
python-vlc==3.0.11115
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
rsa==4.6
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.15.0
typing-extensions==3.7.4.3
typing-inspect==0.6.0
uritemplate==3.0.1
urllib3==1.25.10
vext==0.7.4
vext.gi==0.7.4
Werkzeug==1.0.1
